% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check goodness of fit for 'mobility.model' class}
\usage{
check(object, plots, ...)
}
\arguments{
\item{object}{a \code{mobility.model} object produced by the \code{\link{mobility}} function}

\item{plots}{logical indicating whether to plot the Posterior Predictive Check and Normal Q-Q Plot (default = \code{TRUE})}

\item{...}{further arguments passed to or from other methods}
}
\value{
a list of goodness of fit measures
}
\description{
This function takes a fitted mobility \code{mobility.model} object and calculates goodness of fit metrics. If the Deviance
Information Criterin (DIC) was calculated in the supplied model object, it is included in output.
When \code{plots = TRUE}, two plots are shown containing the posterior distribution of trip counts compared to observed data
and a Normal Q-Q plot showing the quantiles of model residuals against those
expected from a Normal distribution.
}
\details{
Goodness of fit metrics include:
\describe{
  \item{DIC}{\href{https://en.wikipedia.org/wiki/Deviance_information_criterion}{Deviance Information Criterion}}
  \item{RMSE}{\href{https://en.wikipedia.org/wiki/Root-mean-square_deviation}{Root Mean Squared Error}}
  \item{MAPE}{\href{https://en.wikipedia.org/wiki/Mean_absolute_percentage_error}{Mean Absolute Percent Error}}
  \item{R2}{\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{R-squared}}
}
}
\examples{
mod <- mobility(data=mobility_matrices, model='gravity', type='transport', DIC=TRUE)
check(mod)
}
\seealso{
Other model: 
\code{\link{compare}()},
\code{\link{fit_jags}()},
\code{\link{fit_prob_travel}()},
\code{\link{mobility}()},
\code{\link{predict}()},
\code{\link{residuals}()},
\code{\link{summary}()}
}
\author{
John Giles
}
\concept{model}
