% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gravity}
\alias{check_gravity}
\title{Check goodness of fit of a gravity model}
\usage{
check_gravity(M, D, N, mod, plot_check = TRUE)
}
\arguments{
\item{M}{named matrix of trip counts among all \eqn{ij} location pairs}

\item{D}{named matrix of distances among all \eqn{ij} location pairs}

\item{N}{named vector of population sizes for all locations}

\item{mod}{model output from either the \code{\link{fit_gravity}} or \code{\link{summarize_mobility}} functions}

\item{plot_check}{logical indicating whether to plot the Posterior Predictive Check and Normal Q-Q Plot (default = \code{TRUE})}
}
\value{
a list of goodness of fit measures
}
\description{

}
\examples{
M <- mobility_matrices$M
D <- mobility_matrices$D
N <- mobility_matrices$N

mod <- fit_gravity(M, D, N, DIC=TRUE)

check_gravity(M=M, N=N, D=D, mod=mod)
}
\seealso{
Other model: 
\code{\link{fit_gravity}()},
\code{\link{fit_jags}()},
\code{\link{fit_prob_travel}()},
\code{\link{summarize_mobility}()}

Other gravity: 
\code{\link{fit_gravity}()},
\code{\link{sim_gravity}()}
}
\author{
John Giles
}
\concept{gravity}
\concept{model}
