% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_mobility}
\alias{check_mobility}
\title{Check goodness of fit of a mobility model}
\usage{
check_mobility(M, D = NULL, N = NULL, mod, plot_check = TRUE)
}
\arguments{
\item{M}{named matrix of trip counts among all \eqn{ij} location pairs}

\item{D}{named matrix of distances among all \eqn{ij} location pairs}

\item{N}{named vector of population sizes for all locations}

\item{mod}{model output from either the \code{fit_} functions or \code{\link{summarize_mobility}} function}

\item{plot_check}{logical indicating whether to plot the Posterior Predictive Check and Normal Q-Q Plot (default = \code{TRUE})}
}
\value{
a list of goodness of fit measures
}
\description{
This function takes a fitted mobility model and calculates goodness of fit metrics. Model objects produced by the
\code{\link{fit_prob_travel}}, \code{\link{fit_gravity}}, \code{\link{fit_mobility}} or \code{\link{summarize_mobility}}
functions are accepted. If the Deviance Information Criterin (DIC) was calculated in the supplied model object, it is included in output.
When \code{plot_check = TRUE}, two plots are shown containing the posterior distribution of trip counts compared to observed data
and a Normal Q-Q plot showing the quantiles of model residuals against those
expected from a Normal distribution. Goodness of fit metrics include:
\describe{
  \item{DIC}{\href{https://en.wikipedia.org/wiki/Deviance_information_criterion}{Deviance Information Criterion}}
  \item{RMSE}{\href{https://en.wikipedia.org/wiki/Root-mean-square_deviation}{Root Mean Squared Error}}
  \item{MAPE}{\href{https://en.wikipedia.org/wiki/Mean_absolute_percentage_error}{Mean Absolute Percent Error}}
  \item{R2}{\href{https://en.wikipedia.org/wiki/Coefficient_of_determination}{R-squared}}
}
}
\examples{
M <- mobility_matrices$M
D <- mobility_matrices$D
N <- mobility_matrices$N

total <- rowSums(M, na.rm=TRUE)
mod_travel <- summarize_mobility(
  fit_prob_travel(total-diag(M), total)
)

mod_gravity <- summarize_mobility(
  fit_gravity(M, D, N)
)

mod_mobility <- summarize_mobility(
  fit_mobility(M, D, N)
)

check_mobility(M=M,
               mod=mod_travel)

check_mobility(M=M,
               D=D,
               N=N,
               mod=mod_gravity)

check_mobility(M=M,
               D=D,
               N=N,
               mod=mod_mobility)
}
\seealso{
Other model: 
\code{\link{fit_gravity}()},
\code{\link{fit_jags}()},
\code{\link{fit_mobility}()},
\code{\link{fit_prob_travel}()},
\code{\link{summary.mobility.model}()}
}
\author{
John Giles
}
\concept{model}
