% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{combine_rjags}
\alias{combine_rjags}
\title{Combine two rjags objects}
\usage{
combine_rjags(a, b)
}
\arguments{
\item{x}{an rjags list}
}
\value{
an rjags list
}
\description{
This function combines two rjags objects.
}
\author{
John Giles
}
