% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_models.R
\name{fit.gravity}
\alias{fit.gravity}
\title{Fit gravity model to movement matrix}
\usage{
fit.gravity(
  M,
  D,
  N = NULL,
  N.orig = NULL,
  N.dest = NULL,
  n.chain = 2,
  n.burn = 1000,
  n.samp = 1000,
  n.thin = 1,
  prior = NULL,
  DIC = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{M}{named matrix of trip counts among all \eqn{ij} location pairs}

\item{D}{named matrix of distances among all \eqn{ij} location pairs}

\item{N}{named vector of population sizes for all locations (either N or both N.orig and N.dest must be supplied)}

\item{N.orig}{named vector of population sizes for each origin}

\item{N.dest}{named vector of population sizes for each destination.}

\item{n.chain}{number of MCMC sampling chains}

\item{n.burn}{number of iterations to discard before sampling of chains begins (burn in)}

\item{n.samp}{number of iterations to sample each chain}

\item{n.thin}{interval to thin samples}

\item{DIC}{logical indicating whether or not to calculate the Deviance Information Criterion (DIC) (default = FALSE)}

\item{parallel}{logical indicating whether or not to run MCMC chains in parallel or sequentially (default = FALSE)}
}
\value{
a runjags model object containing fitted gravity model paramters
}
\description{
This function fits gravity model parameters to a supplied movement matrix using Bayesian MCMC inference. The function defines the model and serves as a wrapper for the \code{\link{run.jags}}
function in the \code{\link{runjags}} package. Gravity model formula:
\deqn{
    \theta * ( N_i^\omega_1 N_j^\omega_2 / f(d_ij) )
}
}
\seealso{
Other model: 
\code{\link{fit.jags}()},
\code{\link{fit.prob.travel}()},
\code{\link{summarize.hmob}()}

Other gravity: 
\code{\link{fit.prob.travel}()}
}
\author{
John Giles
}
\concept{gravity}
\concept{model}
