% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_models.R
\name{fit.prob.travel}
\alias{fit.prob.travel}
\title{Estimate probability of travelling outside origin}
\usage{
fit.prob.travel(
  V.travel,
  V.tot,
  n.chain = 4,
  n.burn = 1000,
  n.samp = 1000,
  n.thin = 1,
  DIC = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{V.travel}{named vector of total number of people that reported travelling outside their home location}

\item{V.tot}{named vector of the total number of individuals in travel survey for each location}

\item{n.chain}{number of MCMC sampling chains}

\item{n.burn}{number of iterations to discard before sampling of chains begins (burn in)}

\item{n.samp}{number of iterations to sample each chain}

\item{n.thin}{interval to thin samples}

\item{DIC}{logical indicating whether or not to calculate the Deviance Information Criterion (DIC)}

\item{parallel}{logical indicating whether or not to run MCMC chains in parallel or sequentially (default = FALSE)}

\item{n.adapt}{number of adaptive iterations}
}
\value{
dataframe giving input data along with estimates of travel probability for each location
}
\description{
This function fits a hierarchical model that estimates the probability an individual travels outside their home location
within the time period of the survey (tau). The model estimates both the overall population-level probability of travel (tau_pop) and
the origin-level probability of travel (tau_i). Further this method is designed for sparse observations that typically result from
travel survey data, where unobserved routes of travel regress to the population mean.
}
\seealso{
Other model: 
\code{\link{fit.gravity}()},
\code{\link{fit.jags}()},
\code{\link{summarize.hmob}()}

Other gravity: 
\code{\link{fit.gravity}()}
}
\author{
John Giles
}
\concept{gravity}
\concept{model}
