% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_models.R
\name{fit_gravity}
\alias{fit_gravity}
\title{Fit gravity model to movement matrix}
\usage{
fit_gravity(
  M,
  D,
  N = NULL,
  n_orig = NULL,
  n_dest = NULL,
  n_chain = 2,
  n_burn = 1000,
  n_samp = 1000,
  n_thin = 1,
  prior = NULL,
  DIC = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{M}{named matrix of trip counts among all \eqn{ij} location pairs}

\item{D}{named matrix of distances among all \eqn{ij} location pairs}

\item{N}{named vector of population sizes for all locations (either N or both n_orig and n_dest must be supplied)}

\item{n_orig}{named vector of population sizes for each origin}

\item{n_dest}{named vector of population sizes for each destination_}

\item{n_chain}{number of MCMC sampling chains}

\item{n_burn}{number of iterations to discard before sampling of chains begins (burn in)}

\item{n_samp}{number of iterations to sample each chain}

\item{n_thin}{interval to thin samples}

\item{DIC}{logical indicating whether or not to calculate the Deviance Information Criterion (DIC) (default = FALSE)}

\item{parallel}{logical indicating whether or not to run MCMC chains in parallel or sequentially (default = FALSE)}
}
\value{
a runjags model object containing fitted gravity model paramters
}
\description{
This function fits gravity model parameters to a supplied movement matrix using Bayesian MCMC inference. The function defines the model and serves as a wrapper for the \code{\link{run_jags}}
function in the \code{\link{runjags}} package. Gravity model formula:
\deqn{
    \theta * ( N_i^\omega_1 N_j^\omega_2 / f(d_ij) )
}
}
\examples{
n <- 5
ids <- LETTERS[1:n]

# Distance matrix
D <- get_distance_matrix(x=rnorm(n, 100, 5),
                         y=rnorm(n, 20, 2),
                         id=ids)

# Vector of population sizes
N <- rpois(n, 1000)
names(N) <- ids

# Simulate trip counts based on model parameters
M <- sim_gravity(N=N,
                 D=D,
                 theta=20,
                 omega_1=15,
                 omega_2=0.75,
                 gamma=1.5,
                 counts=TRUE)

mod <- fit_gravity(M, D, N)
summarize_mobility(mod)
}
\seealso{
Other model: 
\code{\link{fit_jags}()},
\code{\link{fit_prob_travel}()},
\code{\link{summarize_mobility}()}

Other gravity: 
\code{\link{fit_prob_travel}()},
\code{\link{sim_gravity}()}
}
\author{
John Giles
}
\concept{gravity}
\concept{model}
