% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_mobility}
\alias{fit_mobility}
\title{Fit full mobility model to movement matrix}
\usage{
fit_mobility(
  M,
  D,
  N,
  n_chain = 2,
  n_burn = 1000,
  n_samp = 1000,
  n_thin = 1,
  DIC = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{M}{named matrix of trip counts among all \eqn{ij} location pairs}

\item{D}{named matrix of distances among all \eqn{ij} location pairs}

\item{N}{named vector of population sizes for all locations (either N or both n_orig and n_dest must be supplied)}

\item{n_chain}{number of MCMC sampling chains}

\item{n_burn}{number of iterations to discard before sampling of chains begins (burn in)}

\item{n_samp}{number of iterations to sample each chain}

\item{n_thin}{interval to thin samples}

\item{DIC}{logical indicating whether or not to calculate the Deviance Information Criterion (DIC) (default = \code{FALSE})}

\item{parallel}{logical indicating whether or not to run MCMC chains in parallel or sequentially (default = \code{FALSE})}
}
\value{
a runjags model object containing fitted gravity model paramters
}
\description{
This function fits a full mobility model to the supplied movement matrix using Bayesian MCMC inference. The full mobility model uses
the \code{\link{fit_prob_travel}} function to estimate the probability of travel outside the origin location and the \code{\link{fit_gravity}}
function to estimate travel to all destination locations. Unlike \code{\link{fit_gravity}}, \code{\link{fit_mobility}} requires the diagonal of
movement matrix \code{M} to be filled.
}
\examples{
M <- mobility_matrices$M
D <- mobility_matrices$D
N <- mobility_matrices$N

mod <- fit_mobility(M, D, N)
}
\seealso{
Other model: 
\code{\link{check_mobility}()},
\code{\link{fit_gravity}()},
\code{\link{fit_jags}()},
\code{\link{fit_prob_travel}()},
\code{\link{summary.mobility.model}()}
}
\author{
John Giles
}
\concept{model}
