% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_admin_level}
\alias{get_admin_level}
\title{Find the lowest admin unit}
\usage{
get_admin_level(data)
}
\arguments{
\item{data}{generalized data frame described in \code{\link{travel_data_template}} or derivative thereof}
}
\value{
integer
}
\description{
This function checks which admin levels are in a generalized a data frame formatting like \code{\link{travel_data_sim}} and
returns the lowest admin unit.
}
\examples{
get_admin_level(travel_data_sim)
}
\seealso{
Other utility: 
\code{\link{get_beta_params}()},
\code{\link{get_gamma_params}()},
\code{\link{get_pop_vec}()},
\code{\link{get_unique_coords}()},
\code{\link{get_unique_ids}()}
}
\author{
John Giles
}
\concept{utility}
