% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_crossdist}
\alias{get_crossdist}
\title{Get distance matrix for two different set of coordinates}
\usage{
get_crossdist(xy1, xy2, id1 = NULL, id2 = NULL)
}
\arguments{
\item{xy1}{two column matrix of XY coordinates for first group}

\item{xy2}{two column matrix of XY coordinates for second group}

\item{id1}{optional names for first group}

\item{id2}{optional names for second group}
}
\value{
numeric scalar or vector
}
\description{
Takes XY coordinates of two sets of locations and returns cross distance for all entries.
}
\examples{
# Some XY coords in decimal degrees
xy1 <- data.frame(x=rnorm(3, -90, 1),
                  y=rnorm(3, 30, 1),
                  id=LETTERS[1:3])

xy2 <- data.frame(x=rnorm(5, -90, 1),
                  y=rnorm(5, 35, 1),
                  id=LETTERS[4:8])

D <- get_crossdist(xy1=xy1[,1:2],
                   xy2=xy2[,1:2],
                   id1=xy1[,3],
                   id2=xy2[,3])
D
}
\seealso{
Other data synthesis: 
\code{\link{get_distance_matrix}()},
\code{\link{get_mob_matrix}()},
\code{\link{get_stay_data}()}
}
\author{
John Giles
}
\concept{data synthesis}
