% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mob_matrix}
\alias{get_mob_matrix}
\title{Get mobility matrix from longform data}
\usage{
get_mob_matrix(orig, dest, value, missing_obs = "NA")
}
\arguments{
\item{orig}{vector of origin names}

\item{dest}{vector of destination names}

\item{value}{vector of observed values for each origin/destination pair}

\item{missing_obs}{filler value for missing observations (default = \code{NA})}
}
\value{
named numeric matrix
}
\description{
Takes X and Y coordinates of two locations and returns cross distance for all entries.
}
\examples{
x <- travel_data_sim

x <- cbind(x, get_unique_ids(x, adm_start=2))

M <- get_mob_matrix(orig=x$orig_id,
                    dest=x$dest_id,
                    value=x$trips)

library(ggplot2)
ggplot(data=reshape2::melt(M)) +
  geom_tile(aes(x=factor(destination),
                y=factor(origin),
                fill=value)) +
  xlab('Destination') + ylab("Origin") +
  theme_bw() + theme(axis.text.x=element_text(size=10),
                     axis.text.y=element_text(size=10),
                     axis.title.x=element_text(size=12, margin = margin(t = 15)),
                     axis.title.y=element_text(size=12, margin = margin(r = 15)),
                     legend.position='bottom') +
  viridis::scale_fill_viridis(option='inferno', direction=1) +
  guides(fill=guide_colorbar(title='Observed trips',
                             title.position='top',
                             label.theme=element_text(size=9),
                             barwidth=20,
                             barheight=0.5,
                             frame.colour='black',
                             ticks=TRUE))
}
\seealso{
Other data synthesis: 
\code{\link{get_crossdist}()}
}
\author{
John Giles
}
\concept{data synthesis}
