% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pop_vec}
\alias{get_pop_vec}
\title{Get unique origin and destination names}
\usage{
get_pop_vec(data, orig = TRUE, dest = TRUE)
}
\arguments{
\item{data}{generalized data frame described in \code{\link{travel_data_template}} or derivative thereof}

\item{orig}{logical indicating whether to include origin location in unique names (default= TRUE)}

\item{dest}{logical indicating whether to include destination location in unique names (default= TRUE)}
}
\value{
two column dataframe containing unique names
}
\description{
This function returns coordinates for all unique locations in supplied data.
}
\examples{
x <- cbind(travel_data_sim, get_unique_ids(travel_data_sim, adm_start=2))

# Unique coordinates for all locations
get_pop_vec(x)

# Origins only
get_pop_vec(x, dest=F)

# Destinations only
get_pop_vec(x, orig=F)

# Without supplied ids
get_pop_vec(travel_data_sim)
}
\seealso{
Other utility: 
\code{\link{get_admin_level}()},
\code{\link{get_beta_params}()},
\code{\link{get_gamma_params}()},
\code{\link{get_unique_coords}()},
\code{\link{get_unique_ids}()}
}
\author{
John Giles
}
\concept{utility}
