% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stay_data}
\alias{get_stay_data}
\title{Make data set of travel and stays}
\usage{
get_stay_data(data, data_pred = NULL, agg_adm = NULL)
}
\arguments{
\item{data}{generalized data frame described in \code{\link{travel_data_sim}} or derivative thereof}

\item{data_pred}{generalized data frame containing the admin units at which to predict probability of travel}

\item{agg_adm}{optional argument (logical) giving an arbitarary admin unit over which to aggregate travel/stay counts}
}
\value{
dataframe with same columns as \code{\link{travel_data_sim}} data with columns for counts of travel/stay/total
}
\description{
This function builds a data set containing the number of individuals that remain in their home location (stay) or travel to another location
during the time span of the survey. The output is designed to provide data for the \code{\link{fit_prob_travel}} function. The admin unit used to aggregate the travel/stay
counts depend on the arguemtns supplied:
\enumerate{
\item When \code{data_pred} is supplied and \code{agg_adm = NULL}, the lowest admin unit of \code{data_pred} is used
\item When both \code{data_pred} and \code{agg_adm} are \code{NULL}, the lowest admin unit of \code{data} is used
}
}
\examples{
stay_data <- get_stay_data(travel_data_sim, agg_adm=2)
}
\seealso{
Other data synthesis: 
\code{\link{get_crossdist}()},
\code{\link{get_distance_matrix}()},
\code{\link{get_mob_matrix}()}

Other travel probability: 
\code{\link{fit_prob_travel}()}
}
\author{
John Giles
}
\concept{data synthesis}
\concept{travel probability}
