% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_coords}
\alias{get_unique_coords}
\title{Get unique coordinates}
\usage{
get_unique_coords(data, orig = TRUE, dest = TRUE)
}
\arguments{
\item{data}{generalized data frame described in \code{\link{travel_data_template}} or derivative thereof}

\item{orig}{logical indicating whether to include unique coordinates from origin locations (default=TRUE)}

\item{dest}{logical indicating whether to include unique coordinates from destination locations (default=TRUE)}
}
\value{
three column dataframe containing unique coordinates and location names
}
\description{
This function returns the coordinates for all unique origins and/or destinations in supplied data.
}
\examples{
x <- cbind(travel_data_sim, get_unique_ids(travel_data_sim, adm_start=1))

# Unique coordinates for all locations
get_unique_coords(x)

# Origins only
get_unique_coords(x, dest=FALSE)

# Destinations only
get_unique_coords(x, orig=FALSE)

# Without supplied ids
get_unique_coords(travel_data_sim)
}
\seealso{
Other utility: 
\code{\link{get_admin_level}()},
\code{\link{get_beta_params}()},
\code{\link{get_gamma_params}()},
\code{\link{get_pop_vec}()},
\code{\link{get_unique_ids}()}
}
\author{
John Giles
}
\concept{utility}
