% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unique_ids}
\alias{get_unique_ids}
\title{Get unique origin and destination names}
\usage{
get_unique_ids(
  data,
  orig = TRUE,
  dest = TRUE,
  adm_start = NULL,
  adm_stop = NULL,
  name_class = "character"
)
}
\arguments{
\item{data}{generalized data frame described in \code{\link{travel_data_template}}}

\item{orig}{logical indicating whether to include unique coordinates from origin locations (default=TRUE)}

\item{dest}{logical indicating whether to include unique coordinates from destination locations (default=TRUE)}

\item{adm_start}{highest administrative level to include in unique names (default = NULL, which uses highest observed in the data)}

\item{adm_stop}{lowest administrative level to include in unique names (default = NULL, which uses lowest observed in the data)}

\item{name_class}{character indicating whether unique names should be either a unique character string (\code{name_class = "character"}) or a unique integer code (\code{name_class = "numeric"})}
}
\value{
two column dataframe containing unique names
}
\description{
This function builds unique character string identifiers for all origins and destinations.
}
\examples{
get_unique_ids(travel_data_sim)

# Starting at admin level 1
get_unique_ids(travel_data_sim, adm_start=1)

# Use numeric facto levels
get_unique_ids(travel_data_sim, name_class='numeric')

# Numeric for only the origin locations
get_unique_ids(travel_data_sim, dest=FALSE, name_class='numeric')
}
\seealso{
Other utility: 
\code{\link{get_admin_level}()},
\code{\link{get_beta_params}()},
\code{\link{get_gamma_params}()},
\code{\link{get_pop_vec}()},
\code{\link{get_unique_coords}()}
}
\author{
John Giles
}
\concept{utility}
