% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gravity_priors}
\alias{gravity_priors}
\title{Example priors for gravity model}
\format{dataframe}
\usage{
gravity_priors
}
\description{
This data object is a dataframe containing posterior estimates of the mean and standard deviation of the
four gravity model parameters. The estimates are derived from a gravity model fitted to Call Data Records in Namibia.
}
\examples{
# A list containing shape and rate parameters of Gamma distribution
prior_list <- list(
  theta = get_gamma_params(gravity_priors$Mean[1],
                           gravity_priors$SD[1]),
  omega_1 = get_gamma_params(gravity_priors$Mean[2],
                             gravity_priors$SD[2]),
  omega_2 = get_gamma_params(gravity_priors$Mean[3],
                             gravity_priors$SD[3]*1000),
  gamma = get_gamma_params(gravity_priors$Mean[4],
                           gravity_priors$SD[4]*1000)
)

par(mfrow=c(2,2))
for (i in seq_along(prior_list)) {

  curve(dgamma(x,
               shape=prior_list[[i]][1],
               rate=prior_list[[i]][2]),
        c(0, 0, 0, 1)[i],
        c(100, 100, 1.5, 2)[i],
        ylab='Density',
        xlab=names(prior_list)[i])
}
}
\author{
John Giles
}
\keyword{datasets}
