% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupdist.R
\name{groupdist}
\alias{groupdist}
\title{Group Spatial Points by Distance}
\usage{
groupdist(df, coor = NULL, threshold = NULL, groupvar = NULL)
}
\arguments{
\item{df}{df a data frame object.}

\item{coor}{longitude and latitude of the spatial points in the format of c("lon","lat").}

\item{threshold}{a distance threshold used to determine if points are in the same group. The
format is numeric. Unit is meters.}

\item{groupvar}{grouping object to stratify spatial points. Recommend to be ID for each individual.
If \code{groupvar} is not specified, all spatial points processed based on their original sequence.}
}
\value{
a list of numbers and NAs indicating whether a set of spatial points are in the same group
or not in any groups.
}
\description{
This function groups spatial points based on the distance to the next point. If distance is
smaller than a threshold, points are considered in the same group. Group number for points
outside the threshold is NA. Depending on how many spatial points are in the data frame,
runtime might be long.
}
\examples{
data(mobility)
mobility$distgroup<- groupdist(mobility, coor = c("lon", "lat"), threshold = 100, groupvar = "id")

}
\seealso{
\code{\link{seqgroup}}, \code{\link{slide}}
}
