% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouptime.R
\name{grouptime}
\alias{grouptime}
\title{Group by Time Threshold}
\usage{
grouptime(
  df,
  time = NULL,
  units = c("auto", "secs", "mins", "hours", "days", "weeks"),
  threshold = NULL,
  groupvar = NULL
)
}
\arguments{
\item{df}{a data frame object.}

\item{time}{a POSIXct time object used to calculate time difference}

\item{units}{a character string indicating which units time difference is caluclated in.}

\item{threshold}{a numeric object. Value should have the same unit specified in the \code{units} parameter.}

\item{groupvar}{a character string. The name of a variable that is used to group rows. Time difference is
calculated using the earliest and lastest time in each group. Note: data frame is not sorted in the function,
recommend to sort before running the function.}
}
\value{
a list of 1 and 0 indicating whether a set of rows are in the same group
or not in any groups.
}
\description{
This function groups rows by time difference. A time threshold must be given. If the time difference
is longer than time threshold, rows are put in the same group. Otherwise, group number is NA.
}
\seealso{
\code{\link{seqgroup}}
}
