% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergewithorder.R
\name{mergewithorder}
\alias{mergewithorder}
\title{Merge with Order}
\usage{
mergewithorder(df1, df2, by = NULL)
}
\arguments{
\item{df1}{a data frame object, the order of which needs to be unchanged.}

\item{df2}{a data frame object be joined with \code{df1}}

\item{by}{unique identifiers used to join \code{df1} and \code{df2}, same 
as \code{by} in \code{merge}}
}
\value{
a joined data frame with the same order as \code{df1}
}
\description{
This function merges two dataset without changing the order of the base 
dataset.
}
\examples{
data(mobility)
mobility$distgroup<- groupdist(mobility, coor = c("lon", "lat"), threshold = 100, groupvar = "id")
latmean<- aggregate(lat~distgroup, mobility, mean)
names(latmean)<- c("distgroup","latmean")
mobility<- mergewithorder(mobility, latmean, by = "distgroup")

}
\seealso{
\code{\link{merge}}
}

