% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\docType{data}
\name{mobility}
\alias{mobility}
\title{Mobility Sample Dataset}
\format{A data frame with 3720 rows and 4 variables:
\describe{
  \item{id}{unique identifier of people}
  \item{lon}{longitude of location}
  \item{lat}{latitude of location}
  \item{datetime}{a POSIXct time object of when people are at certain locations}
}}
\usage{
mobility
}
\description{
A dataset contains hypothetical mobility data of two people.
}
\keyword{datasets}

