% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mobility_matrices}
\alias{mobility_matrices}
\title{Example data matrices required by gravity functions}
\format{a list of 3
\describe{
  \item{M}{numeric matrix containing simulated trip counts}
  \item{D}{numeric matrix containing distances amopng locations in decimal degrees}
  \item{N}{numeric vector containing population sizes of each location}
  }}
\usage{
mobility_matrices
}
\description{
This data object is a list containing named matrices required by the \code{\link{sim_gravity}} and \code{\link{fit_gravity}} functions.
The matrices provide hypothetical trip counts representing travel among 10 locations (\eqn{M}), along with distances (\eqn{D}) and population sizes (\eqn{N}).
}
\author{
John Giles
}
\keyword{datasets}
