% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{mobility_region}
\alias{mobility_region}
\title{Helper function for retrieving mobility data for a specific region}
\usage{
mobility_region(region)
}
\arguments{
\item{region}{}
}
\value{

}
\description{
Helper function for retrieving mobility data for a specific region
}
