% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Prediction and simulation method for 'mobility.model' class}
\usage{
predict(object, newdata, nsim, seed, ...)
}
\arguments{
\item{object}{a \code{mobility.model} object containing a fitted mobility model and its associated data}

\item{newdata}{a list containing new data to used in model prediction. If \code{NULL} (the default) the function will simulate the model with the data used for fitting.
\describe{
  \item{\code{D}}{a named matrix of distances between all \eqn{ij} location pairs}
  \item{\code{N}}{a named vector of population sizes for all locations (either N or both N_orig and N_dest must be supplied)}
  \item{\code{N_orig}}{a named vector of population sizes for each origin}
  \item{\code{N_dest}}{a named vector of population sizes for each destination}
}}

\item{nsim}{number of simulations (default = 1).}

\item{seed}{optional integer specifying the call to \code{set.seed} prior to model simulation (default = \code{NULL})}

\item{...}{further arguments passed to or from other methods}
}
\value{
a vector, matrix, or array containing predited or simulated mobility values.
}
\description{
This function uses a fitted \code{mobility.model} object to simulate a connectivity matrix based on estimated parameters.
}
\details{
When \code{nsim = 1}, the prediction matrix is calculated
using the mean point estimate of parameter values. If \code{nsim > 1} then returns and array that contains
\code{nsim} number of simulated replications based on the posterior distributions of each parameter.
}
\examples{
mod <- mobility(data=mobility_matrices,
                model='gravity',
                type='transport')

mod <- mobility(data=mobility_matrices,
                model='departure-diffusion',
                type='power',
                hierarchical=TRUE)

predict(object=mod)

n <- 5
ids <- letters[1:n]

# Distance matrix
D <- get_distance_matrix(x=rnorm(n, -100, 2),
                         y=rnorm(n, 20, 1),
                         id=ids)*111.35

# Vector of population sizes
N <- rnbinom(n, size=5, mu=5000)
names(N) <- ids

# Predict mobility model using new data
predict(object=mod, newdata=list(D=D, N=N))
}
\seealso{
Other model: 
\code{\link{check}()},
\code{\link{compare}()},
\code{\link{fit_jags}()},
\code{\link{fit_prob_travel}()},
\code{\link{mobility}()},
\code{\link{residuals}()},
\code{\link{summary}()}
}
\author{
John Giles
}
\concept{model}
