% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiusofgyration.R
\name{radiusofgyration}
\alias{radiusofgyration}
\title{Calculate Radius of Gyration of Spatial Points}
\usage{
radiusofgyration(df, coor = NULL, time = NULL, time.units = c("hour",
  "date", "month"), groupvar = NULL)
}
\arguments{
\item{df}{a data frame object.}

\item{coor}{longitude and latitude of the spatial points in the format of c("lon","lat").}

\item{time}{a POSIXct time object, used to calculate time period.}

\item{time.units}{time units used to group spatial location points. Choose from hour, date,
and month. Radius of gyration is calculated within each time frame.}

\item{groupvar}{grouping object to stratify time objects. Recommend to be ID for each individual.
If \code{groupvar} is not specified, \code{time.units} will be used to sort data and calculate radius of gyration.}
}
\value{
a list of radius of gyration value matching to each spatial point in data frame. points in
the same time period sepecified in time.units have the same radius of gyration.
}
\description{
This function calculates radius of gyration, defined as the distance moved within a certain
time period. Distance is calculated using Vincenty's formula.
}
\examples{
data("mobility")
mobility$rg<- radiusofgyration(mobility, coor = c("lon","lat"), time = "datetime", time.units = "date", groupvar = "id")

}
\seealso{
\code{\link{sdspatialpoints}}
}

