% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regions.R
\docType{data}
\name{regions}
\alias{regions}
\title{List of geographic regions as 2 letter ISO codes}
\format{
An object of class \code{character} of length 135.
}
\usage{
regions
}
\description{
List of geographic regions as 2 letter ISO codes
}
\keyword{datasets}
