% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{residuals}
\alias{residuals}
\title{Extract model residuals}
\usage{
residuals(object, type, ...)
}
\arguments{
\item{object}{a \code{mobility.model} object produced by the \code{\link{mobility}} function}

\item{type}{the type of residuals to be returned. Available residual types include: \code{'deviance'} (default), \code{'pearson'}, and \code{'raw'}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
a matrix containing model residuals
}
\description{
Generic fundtion that extracts model residuals from a 'mobility.model' object.
}
\details{
Residual types are calculated as:
\describe{
  \item{raw}{\eqn{y_i - \mu_i}}
  \item{pearson}{\eqn{(y_i - \mu_i)/\sqrt{\mu_i}}}
  \item{deviance}{\eqn{sign(y_i - \mu_i) * \sqrt(2(log(y_i/\mu_i) - (y_i - \mu_i)))}}
}
Where, \eqn{y_i} is the observed data and \eqn{\mu_i} is the value predicted by the model using the mean of parameter posterior distributions
}
\examples{
mod <- mobility(data=mobility_matrices, model='gravity', type='transport', DIC=TRUE)

residuals(mod, type='pearson')
}
\seealso{
Other model: 
\code{\link{check}()},
\code{\link{compare}()},
\code{\link{fit_jags}()},
\code{\link{fit_prob_travel}()},
\code{\link{mobility}()},
\code{\link{predict}()},
\code{\link{summary}()}
}
\author{
John Giles
}
\concept{model}
