% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_models.R
\name{rjags_combine}
\alias{rjags_combine}
\title{Combine two rjags objects}
\usage{
rjags_combine(a, b)
}
\arguments{
\item{x}{an rjags list}
}
\value{
an rjags list
}
\description{
This function combines two rjags objects.
}
\author{
John Giles
}
