% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_models.R
\name{rjags_mcmc_list}
\alias{rjags_mcmc_list}
\title{Convert rjags object to mcmc.list}
\usage{
rjags_mcmc_list(x)
}
\arguments{
\item{x}{an rjags list}
}
\value{
an \code{\link[coda:mcmc.list]{mcmc.list}} object
}
\description{
This function converts an rjags object to an mcmc.list.
}
\author{
John Giles
}
