% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{rjags_to_mcmc}
\alias{rjags_to_mcmc}
\title{Convert rjags object to mcmc.list}
\usage{
rjags_to_mcmc(x)
}
\arguments{
\item{x}{an rjags list}
}
\value{
an \code{\link[coda:mcmc.list]{mcmc.list}} object
}
\description{
This function converts an rjags object to an \code{\link[coda:mcmc.list]{mcmc.list}} object
}
\author{
John Giles
}
