% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqgroup.R
\name{seqgroup}
\alias{seqgroup}
\title{Generates Sequential Groups Based on Values}
\usage{
seqgroup(df, var = NULL)
}
\arguments{
\item{df}{a data frame object.}

\item{var}{the variable in \code{df} grouping is based on. Group number changes when
value becomes different with the last value regardless of what the actual value is.}
}
\value{
a numeric list of group numbers
}
\description{
This function groups variable based on its value. The variable is not sorted, so make
sure to sort before running the function. Rows with the same value but not adjacent
are not in the same group.
}
\examples{
A<- rep(1:3, times = 2, each = 2)
A<- data.frame(A)
A$group<- seqgroup(A, var = "A")

}

