% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_gravity}
\alias{sim_gravity}
\title{Simulate connectivity values using gravity model}
\usage{
sim_gravity(
  D,
  N,
  theta = 1,
  omega_1 = 1,
  omega_2 = 1,
  gamma = 1,
  n = 1,
  mod = NULL,
  counts = FALSE
)
}
\arguments{
\item{D}{matrix giving distances among the origins and destinations}

\item{N}{vector of population sizes}

\item{theta}{scalar giving the proportionality constant of gravity formula (default = 1)}

\item{omega_1}{scalar giving exponential scaling of origin population size (default = 1)}

\item{omega_2}{scalar giving exponential scaling of destination population size (default = 1)}

\item{gamma}{scalar giving the dispersal kernel paramater (default = 1)}

\item{n}{number of simulations (requires argument \code{mod} to be supplied with a gravity model object)}

\item{mod}{a gravity model object produced by the \code{\link{fit_gravity}} or \code{\link{summarize_mobility}} functions}

\item{counts}{logical indicating whether or not to return a count variable by scaling the connectivity matrix by origin population size (\eqn{N_i}) (default = FALSE)}
}
\value{
a matrix with values between 0 and 1 (if \code{counts = FALSE}) or positive integers (if \code{counts = TRUE}). If \code{n > 1} then returns and array with 3 dimensions
}
\description{
This function uses the gravity model formula to simulate a connectivity matrix based on the supplied model parameters. The
gravity model formula uses a Gamma distribution as the dispersal kernel in the denominator. A null model (where all model parameters = 1) can be
simulated by supplying only population sizes (\code{N}) and pairwise distances (\code{D}).
}
\examples{
n <- 5
ids <- LETTERS[1:n]

# Distance matrix
D <- get_distance_matrix(x=rnorm(n, 100, 5),
                         y=rnorm(n, 20, 2),
                         id=ids)

# Vector of population sizes
N <- rpois(n, 1000)
names(N) <- ids

# Simulate null model connectivity matrix
pi_hat <- sim_gravity(D=D, N=N)

# Simulate connectivity matrix given gravity model parameters
pi_hat <- sim_gravity(D=D,
                      N=N,
                      theta=14,
                      omega_1=13,
                      omega_2=0.7,
                      gamma=1.5)

# Simulate many realizations of trip counts based on fitted model parameters
M <- mobility_matrices$M
D <- mobility_matrices$D
N <- mobility_matrices$N

mod <- summarize_mobility(
  fit_gravity(M, D, N)
)

M_hat <- sim_gravity(D=D,
                     N=N,
                     n=5,
                     mod=mod,
                     counts=TRUE)
}
\seealso{
Other simulation: 
\code{\link{sim_mobility}()},
\code{\link{sim_param}()},
\code{\link{sim_prob_travel}()}

Other gravity: 
\code{\link{fit_gravity}()}
}
\author{
John Giles
}
\concept{gravity}
\concept{simulation}
