% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_mobility}
\alias{sim_mobility}
\title{Simulate a full mobility model}
\usage{
sim_mobility(
  D,
  N,
  theta = 1,
  omega_1 = 1,
  omega_2 = 1,
  gamma = 1,
  tau = 0.5,
  n = 1,
  mod = NULL,
  counts = FALSE
)
}
\arguments{
\item{D}{matrix giving distances among the origins and destinations}

\item{N}{vector of population sizes}

\item{theta}{scalar giving the proportionality constant of gravity formula (default = 1)}

\item{omega_1}{scalar giving exponential scaling of origin population size (default = 1)}

\item{omega_2}{scalar giving exponential scaling of destination population size (default = 1)}

\item{gamma}{scalar giving the dispersal kernel paramater (default = 1)}

\item{tau}{scalar or vector giving the probability of travel outside origin}

\item{n}{number of simulations (requires argument \code{mod} to be supplied with a gravity model object)}

\item{mod}{a mobility model object produced by the \code{\link{fit_mobility}} or \code{\link{summarize_mobility}} functions}

\item{counts}{logical indicating whether or not to return a count variable by scaling the mobility matrix by origin population size (\eqn{N_i}) (default = FALSE)}
}
\value{
a matrix with values between 0 and 1 (if \code{counts = FALSE}) or positive integers (if \code{counts = TRUE}). If \code{n > 1} then returns and array with 3 dimensions
}
\description{
This function simulates a mobility matrix based on the supplied model parameters. The full mobility model
estimates both quantity of travel within a location (along the matrix diagonal) as well as the quantity
among all origins and destinations (off-diagonals). For multiple stochastic simulations, the \code{mod}
argument must be supplied with a mobility model object produced by the \code{\link{fit_mobility}} function.
If mean values are supplied to the named parameter arguments (\code{theta}, \code{omega_1}, \code{omega_2}, \code{gamma}, \code{tau}),
a single point estimate of the mobility matrix will be returned. A null model (where all model parameters = 1) can be simulated by
supplying only population sizes (\code{N}) and pairwise distances (\code{D}).
}
\examples{
M <- mobility_matrices$M
D <- mobility_matrices$D
N <- mobility_matrices$N

# Simulate null model for given D and N
# assumes 0.5 probability of traveling outside origin and null gravity
# for probability of travel to all destinations
sim_mobility(D, N)

# Simulate point estimates for some mean parameter values
sim_mobility(D, N,
             theta = 3,
             omega_1 = 2,
             omega_2 = 0.7,
             gamma = 1.5,
             tau=rbeta(length(N), 2, 2))

# Simulate multiple stochastic realizations of a mobility matrix from
# fitted mobility model parameters
mod <- summarize_mobility(
  fit_mobility(M, D, N)
)

mod_sim <- sim_mobility(D,
             N,
             n=3,
             mod=mod,
             counts=TRUE)
}
\seealso{
Other simulation: 
\code{\link{sim_gravity}()},
\code{\link{sim_param}()},
\code{\link{sim_prob_travel}()}
}
\author{
John Giles
}
\concept{simulation}
