% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_param}
\alias{sim_param}
\title{Simulate parameter values}
\usage{
sim_param(n, mean, sd, CI_low = NULL, CI_high = NULL)
}
\arguments{
\item{n}{number of simulations}

\item{mean}{mean of parameter posterior distribution}

\item{sd}{standard of parameter posterior distribution}

\item{CI_low}{lower confidence bound of estimated posterior (default = NULL)}

\item{CI_high}{upper confidence bound of estimated posterior (default = NULL)}
}
\value{
a vector of length n
}
\description{
This function takes the summary statistics of the estimated posterior distribution of a parameter
and simulates random values from a normal distribution with its mean and standard deviation. If CI_low and
CI_high are supplied, simulated values are truncated by the confidence bounds.
}
\examples{
M <- mobility_matrices$M
D <- mobility_matrices$D
N <- mobility_matrices$N

mod <- summarize_mobility(
  fit_gravity(M, D, N)
)

param <- 'gamma'
sim_param(n=10,
          mean=mod[param,'Mean'],
          sd=mod[param,'SD'],
          CI_low=mod[param,'HPD2.5'],
          CI_high=mod[param,'HPD97.5'])
}
\seealso{
Other simulation: 
\code{\link{sim_gravity}()},
\code{\link{sim_mobility}()},
\code{\link{sim_prob_travel}()}
}
\author{
John Giles
}
\concept{simulation}
