% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stayevent.R
\name{stayevent}
\alias{stayevent}
\title{Calculate Stay Events}
\usage{
stayevent(df, coor = NULL, time = NULL, dist.threshold = NULL,
  time.threshold = NULL, time.units = c("auto", "secs", "mins", "hours",
  "days", "weeks"), groupvar = NULL, ...)
}
\arguments{
\item{df}{a data frame object.}

\item{coor}{longitude and latitude of the spatial points in the format of c("lon","lat").}

\item{time}{a POSIXct time object, used to calculate time period.}

\item{dist.threshold}{a distance threshold used to determine if points are in the same group. The 
format is numeric. Unit is meters.}

\item{time.threshold}{a numeric object. Value should have the same unit specified in the \code{time.units} 
parameter.}

\item{time.units}{a character string indicating which units time difference is caluclated in.}

\item{groupvar}{grouping object to stratify time objects. Recommend to be ID for each individual. 
If \code{groupvar} is not specified, \code{time.units} will be used to sort data and calculate radius of gyration.}
}
\value{
a data frame with three new columns added showing stayevents and locations.
}
\description{
This function calculates stay events based on spatial locations and time stamps of mobility data. A distance 
threshold and a time threshold are given to determine if points are in the same stay event set. The centroid 
of points in one stay event set is calculated as the spatial location of the stay event.
}
\note{
\code{stayevent} function doesn't sort the data frame. It is recommended to sort data frame based on 
time and then run the function.
}
\examples{
data(mobility)
mobility_stay<- stayevent(mobility, coor = c("lon","lat"), time = "datetime", dist.threshold = 100, 
                 time.threshold = 30, time.units = "mins", groupvar = "id")

}
\references{
Toole, J.L., et al. The path most traveled: Travel demand estimation using big data resources. Transport.
Res. Part C (2015), \link{http://dx.doi.org/10.1016/j.trc.2015.04.022}
}
\seealso{
\code{\link{groupdist}},\code{\link{grouptime}},\code{\link{seqgroup}}
}

