% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility_models.R
\name{summarize.hmob}
\alias{summarize.hmob}
\title{Calculate summary statistics for a model}
\usage{
summarize.hmob(mod, ac.lags = c(2, 5, 10))
}
\arguments{
\item{mod}{an mcmc.list object}

\item{ac.lags}{vector of lags over which to calculate autocorrelation of samples within chains (default = c(2,5,10))}
}
\value{
a dataframe with summary statistics
}
\description{
This is a wrapper function for \code{\link[MCMCvis:MCMCsummary]{MCMCsummary}} that calculates summary statistics
for each parameter in an mcmc.list object. If the model contains deviance and penalty calculations, then
Deviance Information Criterion (DIC) is calculated and appended to the summary.
}
\seealso{
Other model: 
\code{\link{fit.gravity}()},
\code{\link{fit.jags}()},
\code{\link{fit.prob.travel}()}
}
\author{
John Giles
}
\concept{model}
