% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.mobility.model}
\alias{summary.mobility.model}
\title{Calculate summary statistics for a mobility model}
\usage{
\method{summary}{mobility.model}(
  object,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  ac_lags = c(2, 5, 10)
)
}
\arguments{
\item{object}{a \code{mobility.model} object (can also accept a \code{\link[coda:mcmc.list]{mcmc.list}} object)}

\item{probs}{numeric vector giving the quantiles to calculate for each parameter (default = c(0.025, 0.5, 0.975))}

\item{ac_lags}{numeric vector of lags over which to calculate autocorrelation of samples within chains (default = c(2,5,10))}
}
\value{
a dataframe with summary statistics
}
\description{
This is a wrapper function of \code{\link[MCMCvis:MCMCsummary]{MCMCsummary}} that calculates summary statistics for each
parameter in a \code{mobility.model} object. Summary statistics are calculated for all parameters across
each chain along with convergance diagnosics like the Gelman-Rubin convergence diagnostic and (Rhat) and samples
auto-correlation foreach parameter. If the model object contains deviance and penalty parameters, then Deviance Information
Criterion (DIC) is calculated and appended to the summary.
}
\examples{
# Gravity model
M <- mobility_matrices$M
D <- mobility_matrices$D
N <- mobility_matrices$N

mod <- fit_gravity(M, D, N, type='transport', DIC=TRUE)
summary(mod)
}
\seealso{
Other model: 
\code{\link{check_mobility}()},
\code{\link{fit_gravity}()},
\code{\link{fit_jags}()},
\code{\link{fit_prob_travel}()}
}
\author{
John Giles
}
\concept{model}
