% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Calculate summary statistics for 'mobility.model' class}
\usage{
summary(object, probs, ac_lags, ...)
}
\arguments{
\item{object}{a \code{mobility.model} object (can also accept a \code{\link[coda:mcmc.list]{mcmc.list}} object)}

\item{probs}{numeric vector giving the quantiles to calculate for each parameter (default = \code{c(0.025, 0.5, 0.975)})}

\item{ac_lags}{numeric vector of lags over which to calculate autocorrelation of samples within chains (default = \code{c(2,5,10)})}

\item{...}{further arguments passed to or from other methods}
}
\value{
a dataframe with summary statistics
}
\description{
This is a wrapper function of \code{\link[MCMCvis:MCMCsummary]{MCMCsummary}} that calculates summary statistics for each
parameter in a \code{mobility.model} object. Summary statistics are calculated for all parameters across
each chain along with convergance diagnosics like the Gelman-Rubin convergence diagnostic and (Rhat) and samples
auto-correlation foreach parameter. If the model object contains deviance and penalty parameters, then Deviance Information
Criterion (DIC) is calculated and appended to the summary.
}
\examples{
mod <- mobility(data=mobility_matrices, model='gravity', type='transport', DIC=TRUE)
summary(mod)
}
\seealso{
Other model: 
\code{\link{check}()},
\code{\link{compare}()},
\code{\link{fit_jags}()},
\code{\link{fit_prob_travel}()},
\code{\link{mobility}()},
\code{\link{predict}()}
}
\author{
John Giles
}
\concept{model}
