% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_travel_data}
\alias{template_travel_data}
\title{Generalized template for travel data}
\format{a data frame with empty columns and generalized column names
\describe{
  \item{date_start}{date: beginning of the time interval of the survey for each observation}
  \item{date_stop}{date: end of the time interval of the survey for each observation}
  \item{date_span}{integer: time span in days of the survey.}
  \item{indiv_id}{integer: unique individual identifier}
  \item{indiv_age}{numeric: age of participant}
  \item{indiv_sex}{logical: gender of perticipant}
  \item{indiv_type}{character: if individual participants belong to different groups}
  \item{orig_adm0}{character: name of highest administration level of origin location (Country)}
  \item{orig_adm1}{character: name of administration level 1 of origin location (e.g. Division, State)}
  \item{orig_adm2}{character: name of administration level 2 of origin location (e.g. District, County)}
  \item{orig_adm3}{character: name of administration level 3 of origin location (e.g. Sub-district, Province)}
  \item{orig_adm4}{character: name of administration level 4 of origin location (e.g. City, Municipality)}
  \item{orig_adm5}{character: name of administration level 5 of origin location (e.g. Town, Village, Community, Ward)}
  \item{orig_type}{character: administrative type for the origin location (e.g. sub-district, community vs town, or urban vs rural)}
  \item{orig_x}{numeric: longitude of origin location centroid in decimal degrees (centroid of smallest admin unit)}
  \item{orig_y}{numeric: latitude of origin location centroid in decimal degrees (centroid of smallest admin unit)}
  \item{orig_pop}{numeric: population size of lowest administrative unit for origin location}
  \item{dest_adm0}{character: name of highest administration level of destination location (Country)}
  \item{dest_adm1}{character: name of administration level 1 of destination location (e.g. Division, State)}
  \item{dest_adm2}{character: name of administration level 2 of destination location (e.g. District, County)}
  \item{dest_adm3}{character: name of administration level 3 of destination location (e.g. Sub-district, Province)}
  \item{dest_adm4}{character: name of administration level 4 of destination location (e.g. City, Municipality)}
  \item{dest_adm5}{character: name of administration level 5 of destination location (e.g. Town, Village, Community, Ward)}
  \item{dest_type}{character: administrative type for the destination location (e.g. sub-district, community vs town, or urban vs rural)}
  \item{dest_x}{numeric: longitude of destination location in decimal degrees (centroid of smallest admin unit)}
  \item{dest_y}{numeric: latitude of destination location centroid in decimal degrees (centroid of smallest admin unit)}
  \item{dest_pop}{numeric: population size of lowest administrative unit for destination location}
  \item{trips}{numeric: total number of trips individual made from origin to destination during time span of travel survey}
  }}
\usage{
template_travel_data
}
\description{
This template data frame provides a general structure for travel data that integrates with data synthesis and
modeling functions. Stays (indivduals reported as not traveling outside home location) are included in this data frame,
where origin and destination are the same and trips are set to NA. Note that models fitted and then extrapolated
using other data assume that the same method for defining population size is used throughout. Either dates or time span must be filled.
}
\examples{
#--------------------------------
# Travel within the same district
#--------------------------------

trip <- survey.template
n <- 3 # Add some observations
trip[1:n,] <- NA

# Time span of travel survey
trip$date_start <- as.Date("2020-01-01")
trip$date_stop <- trip$date_start + 30
trip$date_span <- difftime(trip$date_stop, trip$date_start, units='days')

# Participant info
trip$indiv_id <- sample(1:100, n)
trip$indiv_age <- round(runif(n, 5, 80))
trip$indiv_sex <- rbinom(n, 1, 0.5)

# Origin info
trip$orig_adm0 <- 'A'
trip$orig_adm1 <- 'A'
trip$orig_adm2 <- 'A'
trip$orig_adm3 <- LETTERS[1:n]
trip$orig_type <- 'Sub-district' # Type of admin unit for lowest admin level
trip$orig_x <- rnorm(n, 100, 5)
trip$orig_y <- rnorm(n, 20, 2)
trip$orig_pop <- rpois(n, 10000)

# Destination info
trip$dest_adm0 <- 'A'
trip$dest_adm1 <- 'A'
trip$dest_adm2 <- 'B'
trip$dest_adm3 <- LETTERS[(n+1):(n*2)]
trip$dest_type <- 'Sub-district' # Type of admin unit for lowest admin level
trip$dest_x <- rnorm(n, 100, 5)
trip$dest_y <- rnorm(n, 20, 2)
trip$dest_pop <- rpois(n, 5000)

# Number of reported trips
trip$trips <- rpois(n, 10)

head(trip)



#-----------------------
# Stays in home location
#-----------------------

stay <- survey.template
n <- 3 # add some observations
stay[1:n,] <- NA

# Time span of travel survey
stay$date_start <- as.Date("2020-01-01")
stay$date_stop <- stay$date_start + 30
stay$date_span <- difftime(trip$date_stop, trip$date_start, units='days')

# Participant info
stay$indiv_id <- sample(100:200, n)
stay$indiv_age <- round(runif(n, 5, 80))
stay$indiv_sex <- rbinom(n, 1, 0.5)

# Origin info
stay$orig_adm0 <- stay$dest_adm0 <- 'A'
stay$orig_adm1 <- stay$dest_adm1 <- 'A'
stay$orig_adm2 <- stay$dest_adm2 <- 'A'
stay$orig_adm3 <- stay$dest_adm3 <- LETTERS[1:n]
stay$orig_type <- stay$dest_type <- 'Sub-district'
stay$orig_x <- stay$dest_x <- rnorm(n, 100, 5)
stay$orig_y <- stay$dest_y <- rnorm(n, 20, 2)
stay$orig_pop <- stay$dest_pop <- rpois(n, 10000)

stay$trips <- NA

head(stay)

# Combine
survey_data <- dplyr::full_join(trip, stay)
head(survey_data)



#----------------------------------------
# Dataset with which to extrapolate model
#----------------------------------------

pred <- survey.template
n <- 6 # Add some observations
pred[1:n,] <- NA

# Time span of the interval over which to extrapolate the fitted model
pred$date_span <- as.difftime(7, units='days')

# Origin info
pred$orig_adm0 <- 'A'
pred$orig_adm1 <- 'A'
pred$orig_adm2 <- LETTERS[1:n]
pred$orig_type <- 'District' # Type of admin unit for lowest admin level
pred$orig_x <- rnorm(n, 100, 5)
pred$orig_y <- rnorm(n, 20, 2)
pred$orig_pop <- rpois(n, 1e+05)

# Number of reported trips (unobserved for extrapolation data)
trip$trips <- NA 

head(pred)
}
\author{
John Giles
}
\keyword{datasets}
