% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{travel_data_sim}
\alias{travel_data_sim}
\title{Simulated travel data}
\format{a \code{\link{travel_data_template}} data frame}
\usage{
travel_data_sim
}
\description{
This data set contains simulated values of location information and observed number of trips among locations and within
home locations. Data structure inherited from the \code{\link{travel_data_template}} object.
}
\author{
John Giles
}
\keyword{datasets}
