% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{combine_rjags}
\alias{combine_rjags}
\title{Combine two rjags objects}
\usage{
combine_rjags(a, b)
}
\arguments{
\item{a}{an rjags list}

\item{b}{an rjags list}
}
\value{
an rjags list containing model ouput from \code{a} and \code{b}
}
\description{
This function combines two rjags model objects.
}
\author{
John Giles
}
