% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_prob_travel}
\alias{fit_prob_travel}
\title{Estimate probability of travelling outside origin}
\usage{
fit_prob_travel(
  travel,
  total,
  n_chain = 4,
  n_burn = 1000,
  n_samp = 1000,
  n_thin = 1,
  DIC = FALSE,
  parallel = FALSE,
  format_locations = FALSE
)
}
\arguments{
\item{travel}{named vector of number of people that reported travelling outside their home location}

\item{total}{named vector of the total number of individuals in travel survey for each location}

\item{n_chain}{number of MCMC sampling chains}

\item{n_burn}{number of iterations to discard before sampling of chains begins (burn in)}

\item{n_samp}{number of iterations to sample each chain}

\item{n_thin}{interval to thin samples}

\item{DIC}{logical indicating whether or not to calculate the Deviance Information Criterion (DIC)}

\item{parallel}{logical indicating whether or not to run MCMC chains in parallel or sequentially (default = \code{FALSE})}

\item{format_locations}{logical indicating which format to return results:
\describe{
  \item{\code{FALSE}}{returns only summary of estimated parameters (default)}
  \item{\code{TRUE}}{returns probability of travel with input data, including missing locations}
}}
}
\value{
dataframe giving input data along with estimates of travel probability for each location
}
\description{
This function fits a hierarchical beta-binomial model that estimates the probability an individual travels outside their home location
within the time period of the survey (tau). The model estimates both the overall population-level probability of travel (tau_pop) and
the origin-level probability of travel (tau_i). Further this method is designed for sparse observations that typically result from
data with high temporal resolution or travel survey data. When data are missing, unobserved routes of travel regress to the population mean.
}
\examples{
n_orig <- 6
n_missing <- 3
orig_id <- LETTERS[1:n_orig]

N <- rpois(n_orig, 100)    # population size of each origin
p <- rbeta(n_orig, 2, 2)   # probability of leaving origin

V_travel <- setNames(rbinom(n_orig, N, p), orig_id)
V_tot <- setNames(N, orig_id)

miss <- sample(1:n_orig, n_missing) # missing observations
V_travel[miss] <- V_tot[miss] <- NA

# Estimate probability of travel for observed locations and population mean
prob_trav <- fit_prob_travel(travel=V_travel,
                             total=V_tot,
                             DIC=TRUE)
prob_trav

# Estimate probability of travel for each locations (missing locations regress to mean)
prob_trav <- fit_prob_travel(travel=V_travel,
                             total=V_tot,
                             format_locations=TRUE)
prob_trav


library(ggplot2)
ggplot(data=prob_trav) +
  geom_vline(aes(xintercept=Mean[ which(is.na(travel))[1] ]), color='red', linetype=2) +
  geom_point(aes(x=Mean, y=orig_id), size=2) +
  ggstance::geom_linerangeh(aes(y=orig_id, xmin=CI2.5, xmax=CI97.5)) +
  xlab('Probability of travel outside origin') + ylab('Origin') +
  theme_bw() + theme(axis.text.x=element_text(size=8),
                     axis.text.y=element_text(size=6),
                     axis.title.x=element_text(size=12, margin = margin(t = 15)),
                     axis.title.y=element_text(size=12, margin = margin(r = 15)),
                     legend.position='right')
}
\seealso{
Other model: 
\code{\link{fit_gravity}()},
\code{\link{fit_jags}()}
}
\author{
John Giles
}
\concept{model}
