% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_distance_matrix}
\alias{get_distance_matrix}
\title{Build distance matrix from XY coordinates}
\usage{
get_distance_matrix(x, y, id)
}
\arguments{
\item{x}{vector giving X coordinates}

\item{y}{vector giving Y coordinates}

\item{id}{vector of names for each location}
}
\value{
a named matrix of pairwise distances among locations
}
\description{
This function builds the pairwise distance matrix from vectors of XY coordinates and associated names.
}
\examples{
n <- 5

xy <- data.frame(x=rnorm(n, 100, 1),
                 y=rnorm(n, 20, 1),
                 id=LETTERS[1:n])

D <- get_distance_matrix(x=xy[,1],
                         y=xy[,2],
                         id=xy[,3])
}
\author{
John Giles
}
\concept{data synthesis}
