% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{mobility}
\alias{mobility}
\title{Fit mobility model to data}
\usage{
mobility(
  data,
  model,
  type,
  hierarchical = FALSE,
  n_chain = 2,
  n_burn = 1000,
  n_samp = 1000,
  n_thin = 1,
  DIC = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{data}{a list containing mobility data and covariates. The list object must include EITHER \code{M}, \code{D}, and \code{N}
OR \code{M}, \code{D}, \code{N_orig}, and \code{N_dest}. Details about each item are as follows:
\describe{
  \item{M}{named matrix of trip counts between all \eqn{ij} location pairs}
  \item{D}{named matrix of distances between all \eqn{ij} location pairs}
  \item{N}{named vector of population sizes for all locations (either N or both N_orig and N_dest must be supplied)}
  \item{N_orig}{named vector of population sizes for each origin}
  \item{N_dest}{named vector of population sizes for each destination}
}}

\item{model}{character vector indicating which mobility model to fit to the data. Mobility models include: \code{'gravity'},
\code{'radiation'}, and \code{'departure-diffusion'}.}

\item{type}{character vector indicating the particular sub-type of mobility model to fit.
\describe{
  \item{Gravity model types include:}{\code{basic}, \code{transport}, \code{power}, \code{exp}, \code{power_norm}, \code{exp_norm}, and \code{Marshall}}
  \item{Radiation model types include:}{\code{basic} and \code{finite}}
  \item{Departure-diffusion model types include:}{\code{power}, \code{exp}, and \code{radiation}}
  }
See model list vignette for more detailed description of each model type.}

\item{hierarchical}{Applies only to the \code{'departure-diffusion'} model. A logical argument indicating whether or not to fit \eqn{\tau}
as a single parameter or hierarchically (as \eqn{\tau_{pop}} and \eqn{\tau_i}).}

\item{n_chain}{number of MCMC sampling chains}

\item{n_burn}{number of iterations to discard before sampling of chains begins (burn in)}

\item{n_samp}{number of iterations to sample each chain}

\item{n_thin}{interval to thin samples}

\item{DIC}{logical indicating whether or not to calculate the Deviance Information Criterion (DIC) (default = \code{FALSE})}

\item{parallel}{logical indicating whether or not to run MCMC chains in parallel or sequentially (default = \code{FALSE})}
}
\value{
An object of class \code{mobility.model} containing model information, data, and fitted model parameters
}
\description{
This function fits a variety of mobility models to a supplied movement matrix (\code{M}) and covariates (\code{D} and \code{N}) using Bayesian MCMC inference.
The function specifies the type of mobility model and serves as a wrapper for the \code{\link{fit_jags}} function.
}
\examples{
mod <- mobility(data=mobility_matrices, model='gravity', type='transport')

mod <- mobility(data=mobility_matrices, model='radiation', type='finite')

mod <- mobility(data=mobility_matrices, model='departure-diffusion', type='power')
}
\seealso{
Other model: 
\code{\link{check}()},
\code{\link{compare}()},
\code{\link{fit_jags}()},
\code{\link{fit_prob_travel}()},
\code{\link{predict}()},
\code{\link{summary}()}
}
\author{
John Giles
}
\concept{model}
