% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_gravity}
\alias{sim_gravity}
\title{Simulate connectivity values using gravity model}
\usage{
sim_gravity(
  N,
  D,
  theta = 1,
  omega_1 = 1,
  omega_2 = 1,
  gamma = 1,
  counts = FALSE
)
}
\arguments{
\item{N}{vector of population sizes}

\item{D}{matrix giving distances among the origins and destinations}

\item{theta}{scalar giving the proportionality constant of gravity formula (default = 1)}

\item{omega_1}{scalar giving exponential scaling of origin population size (default = 1)}

\item{omega_2}{scalar giving exponential scaling of destination population size (default = 1)}

\item{gamma}{scalar giving the dispersal kernel paramater (default = 1)}

\item{counts}{logical indicating whether or not to return a count variable by scaling the connectivity matrix by origin population size (\eqn{N_i}) (default = FALSE)}
}
\value{
a matrix with values between 0 and 1 (if \code{counts = FALSE}) or positive integers (if \code{counts = TRUE})
}
\description{
This function uses the gravity model formula to simulate a connectivity matrix based on the supplied model parameters. The
gravity model formula uses a Gamma distribution as the dispersal kernel in the denominator. A null model (where all model parameters = 1) can be
simulated by supplying only population sizes (\code{N}) and pairwise distances (\code{D}).
}
\examples{
n <- 5
ids <- LETTERS[1:n]

# Distance matrix
D <- get_distance_matrix(x=rnorm(n, 100, 5),
                         y=rnorm(n, 20, 2),
                         id=ids)

# Vector of population sizes
N <- rpois(n, 1000)
names(N) <- ids

# Simulate null model connectivity matrix
pi_hat <- sim_gravity(N=N, D=D)

# Simulate connectivity matrix given gravity model parameters
pi_hat <- sim_gravity(N=N,
                      D=D,
                      theta=14,
                      omega_1=13,
                      omega_2=0.7,
                      gamma=1.5)

# Simulate trip counts based on fitted model parameters
M <- mobility_matrices$M
D <- mobility_matrices$D
N <- mobility_matrices$N

mod <- summarize_mobility(
  fit_gravity(M, D, N)
)

M_hat <- sim_gravity(N=N,
                     D=D,
                     theta=mod['theta', 'Mean'],
                     omega_1=mod['omega_1', 'Mean'],
                     omega_2=mod['omega_2', 'Mean'],
                     gamma=mod['gamma', 'Mean'],
                     counts=TRUE)
}
\seealso{
Other simulation: 
\code{\link{get_gamma_params}()},
\code{\link{sim_prob_travel}()}

Other gravity: 
\code{\link{check_gravity}()},
\code{\link{fit_gravity}()}
}
\author{
John Giles
}
\concept{gravity}
\concept{simulation}
