% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sim_prob_travel}
\alias{sim_prob_travel}
\title{Simulate probability of leaving origin}
\usage{
sim_prob_travel(mu = 0.5, sigma = 0.1, n = 1, id = NULL)
}
\arguments{
\item{mu}{scalar or vector giving mean probability of leaving origin (defualt = 0.5)}

\item{sigma}{scalar or vector giving standard deviation of the probability of leaving origin (default = 0.1)}

\item{n}{number of simulations (default = 1)}

\item{id}{optional scalar or vector giving name(s) of origin (default = \code{NULL})}
}
\value{
a numeric scalar or vector with values between 0 and 1
}
\description{
This function simulates one stochastic realization of the proportion of individuals that leave origin \eqn{i}.
The function takes posterior estimates of the mean (\code{mu}) and standard deviation (\code{sigma})
for each origin \eqn{i}, and then derives the shape and rate parameters for the Beta distribution. Simulated values are random
draws from this Beta distribution.
}
\examples{
sim_prob_travel(0.5, 0.04)

sim_prob_travel(mu=c(0.8, 0.6),
                sigma=c(0.08, 0.06),
                id=c('A', 'B'))

# Simulate with estimated parameters
n_orig <- 6
n_missing <- 3
orig_id <- LETTERS[1:n_orig]

N <- rpois(n_orig, 100)    # population size of each origin
p <- rbeta(n_orig, 2, 2)   # probability of leaving origin

travel <- setNames(rbinom(n_orig, N, p), orig_id)
total <- setNames(N, orig_id)

miss <- sample(1:n_orig, n_missing) # missing observations
travel[miss] <- total[miss] <- NA

# Estimate probability of travel for each locations (missing locations regress to mean)
prob_trav <- summarize_mobility(
  fit_prob_travel(travel=travel, total=total)
)

sim_prob_travel(mu=prob_trav$Mean,
                sigma=prob_trav$SD,
                id=names(travel),
                n=5)
}
\seealso{
Other simulation: 
\code{\link{sim_gravity}()},
\code{\link{sim_mobility}()},
\code{\link{sim_param}()}

Other travel probability: 
\code{\link{fit_prob_travel}()},
\code{\link{get_stay_data}()}
}
\author{
John Giles
}
\concept{simulation}
\concept{travel probability}
