% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.mobility.model}
\alias{summary.mobility.model}
\title{Calculate summary statistics for a mobility model}
\usage{
\method{summary}{mobility.model}(object, probs = c(0.025, 0.5, 0.975), ac_lags = c(2, 5, 10))
}
\arguments{
\item{object}{a \code{mobility.model} object (can also accept a \code{\link[coda:mcmc.list]{mcmc.list}} object)}

\item{probs}{numeric vector giving the quantiles to calculate for each parameter (default = c(0.025, 0.5, 0.975))}

\item{ac_lags}{numeric vector of lags over which to calculate autocorrelation of samples within chains (default = c(2,5,10))}
}
\value{
a dataframe with summary statistics
}
\description{
This is a wrapper function of \code{\link[MCMCvis:MCMCsummary]{MCMCsummary}} that calculates summary statistics for each
parameter in a \code{mobility.model} object. Summary statistics are calculated for all parameters across
each chain along with convergance diagnosics like the Gelman-Rubin convergence diagnostic and (Rhat) and samples
auto-correlation foreach parameter. If the model object contains deviance and penalty parameters, then Deviance Information
Criterion (DIC) is calculated and appended to the summary.
}
\examples{
# Gravity model
M <- mobility_matrices$M
D <- mobility_matrices$D
N <- mobility_matrices$N

mod <- fit_gravity(M, D, N, DIC=TRUE)
summary(mod)


# Probability of travel
n_orig <- 10
n_missing <- 3
orig_id <- LETTERS[1:n_orig]

N <- rpois(n_orig, 100)    # population size of each origin
p <- rbeta(n_orig, 1, 2)   # probability of leaving origin

travel <- setNames(rbinom(n_orig, N, p), orig_id)
total <- setNames(N, orig_id)

miss <- sample(1:n_orig, n_missing) # missing observations
travel[miss] <- total[miss] <- NA

# Estimate probability of travel for each locations (missing locations regress to mean)
prob_trav <- summarize_mobility(
  fit_prob_travel(travel=travel, total=total)
)

library(ggplot2)
ggplot(data=prob_trav) +
  geom_point(aes(x=Mean, y=orig_id), size=2) +
  ggstance::geom_linerangeh(aes(y=orig_id, xmin=HPD2.5, xmax=HPD97.5)) +
  xlab('Probability of travel outside origin') + ylab('Origin') +
  xlim(0,1) +
  theme_bw() + theme(axis.text.x=element_text(size=10),
                     axis.text.y=element_text(size=10),
                     axis.title.x=element_text(size=12, margin = margin(t = 15)),
                     axis.title.y=element_text(size=12, margin = margin(r = 15)),
                     panel.border = element_rect(colour = "black", fill=NA, size=1),
                     legend.position='right')
}
\seealso{
Other model: 
\code{\link{check_mobility}()},
\code{\link{fit_gravity}()},
\code{\link{fit_jags}()},
\code{\link{fit_mobility}()},
\code{\link{fit_prob_travel}()}
}
\author{
John Giles
}
\concept{model}
